requirements.txt
openai
python-dotenv
pymongo
scikit-learn
numpy
pandas
PyPDF2
python-docx
reportlab
streamlit